Split wptrunner formatters into separate files.
We have 2 formatters in the same file right now, and a third one to
be added. Splitting wptreport and wptscreenshot pre-emptively so
the new formatter can go its own file in this directory as well.
