+++ tools/wptrunner/wptrunner/formatters/wptreport.py
@@ -3,7 +3,7 @@ import re
 import sys
 
 from mozlog.structured.formatters.base import BaseFormatter
-from .executors.base import strip_server
+from ..executors.base import strip_server
 
 
 LONE_SURROGATE_RE = re.compile(u"[\uD800-\uDFFF]")
@@ -161,23 +161,3 @@ class WptreportFormatter(BaseFormatter):
                                     "process": data["process"]})
 
 
-class WptscreenshotFormatter(BaseFormatter):
-    """Formatter that outputs screenshots in the format expected by wpt.fyi."""
-
-    def __init__(self):
-        self.cache = set()
-
-    def suite_start(self, data):
-        # TODO: ask wpt.fyi for known hashes.
-        pass
-
-    def test_end(self, data):
-        if "reftest_screenshots" not in data.get("extra", {}):
-            return
-        output = ""
-        for item in data["extra"]["reftest_screenshots"]:
-            if type(item) != dict or item["hash"] in self.cache:
-                continue
-            self.cache.add(item["hash"])
-            output += "data:image/png;base64,{}\n".format(item["screenshot"])
-        return output if output else None
